# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TACHIKOMA-OS Backend API",
    "description": "Central orchestrator API for the TACHIKOMA-OS AI ecosystem. Provides LLM gateway, GraphRAG memory, chat, agent tools, and data layer for microservices.",
    "contact": {
      "name": "TACHIKOMA-OS Team"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agent/execute": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "POST /api/agent/execute",
        "operationId": "execute_command",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Command executed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Execution error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Command not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Check error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/search": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "POST /api/agent/search",
        "operationId": "web_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebSearchResultDto"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Search error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandExecuteRequest": {
        "type": "object",
        "description": "Command execution request",
        "required": [
          "command"
        ],
        "properties": {
          "command": {
            "type": "string",
            "description": "Command to execute"
          },
          "timeout_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "Optional timeout in seconds",
            "nullable": true,
            "minimum": 0
          },
          "working_directory": {
            "type": "string",
            "description": "Optional working directory",
            "nullable": true
          }
        }
      },
      "CommandResultDto": {
        "type": "object",
        "description": "Command execution result",
        "required": [
          "exit_code",
          "stdout",
          "stderr",
          "duration_ms",
          "timed_out"
        ],
        "properties": {
          "duration_ms": {
            "type": "integer",
            "format": "int64",
            "description": "Execution duration in ms",
            "minimum": 0
          },
          "exit_code": {
            "type": "integer",
            "format": "int32",
            "description": "Exit code"
          },
          "stderr": {
            "type": "string",
            "description": "Standard error"
          },
          "stdout": {
            "type": "string",
            "description": "Standard output"
          },
          "timed_out": {
            "type": "boolean",
            "description": "Whether command timed out"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "API error response",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "description": "Optional details",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "WebSearchRequest": {
        "type": "object",
        "description": "=============================================================================\nAgent DTOs\n=============================================================================\nWeb search request",
        "required": [
          "query"
        ],
        "properties": {
          "category": {
            "type": "string",
            "description": "Category filter",
            "nullable": true
          },
          "limit": {
            "type": "integer",
            "description": "Maximum results",
            "minimum": 0
          },
          "query": {
            "type": "string",
            "description": "Search query"
          }
        }
      },
      "WebSearchResultDto": {
        "type": "object",
        "description": "Web search result",
        "required": [
          "title",
          "url",
          "snippet",
          "source"
        ],
        "properties": {
          "score": {
            "type": "number",
            "format": "double",
            "description": "Optional score",
            "nullable": true
          },
          "snippet": {
            "type": "string",
            "description": "Result snippet"
          },
          "source": {
            "type": "string",
            "description": "Source engine"
          },
          "title": {
            "type": "string",
            "description": "Result title"
          },
          "url": {
            "type": "string",
            "description": "Result URL"
          }
        }
      }
    }
  },
...
}